feat/running: Finish on Password manger & Implement Running Section.

+++ actions/delete/chatai.ts
@@ -2,13 +2,20 @@
 
 import { prisma } from "@/lib/prisma";
 import { revalidatePath } from "next/cache";
+import { getConnectedUser } from "../auth/user";
 
 export async function deleteChatais() {
   try {
-    const chatais = await prisma.chatai.deleteMany();
+    const { user } = await getConnectedUser();
+
+    if (!user) {
+      throw new Error("Identification inconnu");
+    }
+
+    await prisma.chatai.deleteMany();
 
     revalidatePath("/");
-    return { message: "Chatais deleted successfully", chatais };
+    return { success: true, message: "Chats supprimés" };
   } catch (err) {
     if (err instanceof SyntaxError) {
       return { error: err.message as string };
