+++ actions/delete/rdv.ts
@@ -2,17 +2,24 @@
 
 import { prisma } from "@/lib/prisma";
 import { revalidatePath } from "next/cache";
+import { getConnectedUser } from "../auth/user";
 
 export async function deleteRdv(rdvId: string) {
   try {
-    const rdv = await prisma.rdv.delete({
+    const { user } = await getConnectedUser();
+
+    if (!user) {
+      throw new Error("Identification inconnu");
+    }
+
+    await prisma.rdv.delete({
       where: {
         id: rdvId,
       },
     });
 
     revalidatePath("/");
-    return { message: "RDV deleted successfully", rdv };
+    return { success: true, message: "RDV supprimé" };
   } catch (err) {
     if (err instanceof SyntaxError) {
       return { error: err.message as string };
